refactor: rename service
The `AddUserService` is responsible for logging in users into a npm registry.

The service and associated symbols were renamed to `NpmLoginService`  to better reflect this.
